Extract SizeConstrainer from the IIIFResource
The IIIFResource is a controller, but the SizeConstrainer is a model
